% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_config_Dotto.R
\name{build_site_config}
\alias{build_site_config}
\title{Create a site config file}
\usage{
build_site_config()
}
\description{
It generates a \code{Motto.js} file that comprises all Dotto config files
and save it in the \code{assets} folder.
}
