% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_build_site_DataMotto.R
\name{build_site_DataMotto}
\alias{build_site_DataMotto}
\title{Build Data Motto Website}
\usage{
build_site_DataMotto(open = T)
}
\arguments{
\item{open}{A logical value that opens the Index page of site in the browser. Default is \code{TRUE}.}
}
\description{
Build Data Motto Website
}
\details{
It re-generate a \code{posts.json}, update the \code{index.html}, and create the whole
website based on the updated \code{posts} folder.
}
