% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_create_Dotto.R
\name{create_Dotto}
\alias{create_Dotto}
\title{Create a new Dotto}
\usage{
create_Dotto(title = NULL, open = T)
}
\arguments{
\item{title}{Title of the dotto post.}

\item{open}{If open the dotto file in Rstudio or not. Default is \code{TRUE}.}
}
\description{
It creates a dotto based on a default \code{slug} created automatically based on title.
}
