% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{create_slug}
\alias{create_slug}
\title{Create a valid slug from a text}
\usage{
create_slug(title, slug = NULL)
}
\description{
This function modify the text as follows:
\itemize{
\item change to lowecase
\item remove any characters outside a-z A-Z 0-9 and -
\item remove multile hyphens
\item trim hyphen from the start and end of text
@note Idea from rstudio/distill
}
}
