% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_build_site_DataMotto.R
\name{run_root_rmds_if_needed}
\alias{run_root_rmds_if_needed}
\title{Runs all \verb{_} Rmd files in the root path of Data Motto if the updated date of html is behind.}
\usage{
run_root_rmds_if_needed(path)
}
\description{
@description This function compares the \verb{_} Rmd and \code{html} created time. If the html time is behind,
it runs the Rmd (and deleted the \verb{_} from the beginning of the name).
}
