% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_update_Dotto.R
\name{update_Dotto}
\alias{update_Dotto}
\title{Update Dotto file and folder names based on updated \code{slug}}
\usage{
update_Dotto(path = NULL, clean_dotto_deps = T)
}
\arguments{
\item{path}{a path to the Dotto \code{.Rmd} file. IF the path is a rmd folder path, it automatically find the Rmd path
if there is only one unique Rmd file in the folder.}

\item{clean_Dotto_deps}{Delete all the files and folders in Dotto and only keep .Rmd file. Default is \code{FALSE}.}
}
\description{
The Dotto url is created based on the name on the Dotto folder, which
follows the format \code{D#-date-slug}, e.g. \code{D112-2020-11-21-data-sampling}. As the Dotto
file and folder names are created originally by \code{slug}, for changing the file/folder
names, it is recommended to change \code{slug} first, and then run \code{update_Dotto()}.
}
\details{
for creating a valid \code{slug}, you can use \code{DataMotto::create_slug("foo")}
}
