% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_upgrade_site_DataMotto.R
\name{upgrade_site_DataMotto}
\alias{upgrade_site_DataMotto}
\title{Upgrading Website by adding a new Dotto}
\usage{
upgrade_site_DataMotto(path = NULL, keep_original_site_libs = T, open = T)
}
\arguments{
\item{path}{The path into the Dotto \code{.Rmd}. If the path directs to the Dotto folder
and there exists only one \code{.Rmd} there, it automatically update the path to
select the Dotto Rmd file. Default is \code{NULL} that selects the current working directory.}

\item{keep_original_site_libs}{a logical value that keeps the original \code{site_libs} folder, created by \code{\link{build_site_DataMotto}}
rather than adding a new \code{site_libs} created by rendering \code{Index.Rmd}. Default is \code{TRUE}.}

\item{open}{A logical value that opens the Index page of site in the browser. Default is \code{TRUE}.}
}
\description{
It renders the new Dotto and adds it into the Data Motto, which was already created
and then renders \code{Index.Rmd} to upgrade the cards in the main page.
}
