% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_use_Dotto.R
\name{use_Dotto}
\alias{use_Dotto}
\title{Dotto extensions}
\usage{
use_Dotto(metadata)
}
\description{
Adds required extensions that are necessary for setting up
Rmd file gets \code{Dotto} style.
}
\details{
Add the following code chunk to your Rmarkdown file with output \code{DataMotto::Dotto}.
\code{```{r datamotto_dotto, echo=FALSE}
        DataMotto::use_Dotto()
        ```}
}
\examples{
\dontrun{
DataMotto::use_Dotto()
}
}
